requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.7
django-cleanup==4.0.0
django-cors-headers==3.2.0
django-extensions==2.2.5
django-filter==2.2.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-seed==0.1.9
djangorestframework==3.10.3
Faker==3.0.0
gunicorn==20.0.4
idna==2.8
Pillow==6.2.1
psycopg2==2.8.4
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.7
whitenoise==4.1.4
